Add a delayed reload after submitting modals
Because of the delay in getting new blockchain data to the ledger sync,
After a user submits a modal, the next page will initially load without
the new data. The proper solution to this would be a websocket or some
other sort of notification system. The quick fix is to wait two seconds
and then reload the page.

Signed-off-by: Zac Delventhal <zac@bitwise.io>